+++ src/lib/types/index.ts
@@ -1,4 +1,5 @@
 export * from "./support.types";
 export * from "./common/chip.types";
-export * from "./mover.types";
+export * from "./profile.types";
 export * from "./request.types";
+export * from "./basicInfo.types";
