+++ src/lib/types/profile.types.ts
@@ -1,5 +1,3 @@
-import React from "react";
-
 //기사님 프로필 컴포넌트 모음집
 export interface InputFieldProps {
    name: string; //useActionsState로 매핑하기 위한
@@ -14,23 +12,14 @@ export interface InputFieldProps {
    validator?: (value: string | string[]) => ValidationResult; // 유효성 검사 함수
 }
 
+//기사님 프로필 유효성 함수 관련
 export type ValidationResult = {
    success: boolean;
    message: string;
 };
 
+//기사님 프로필 관련
 export type profileState = {
    status: boolean;
    error?: string;
 } | null;
-
-//기사님 기본정보 컴포넌트 모음집
-export interface BasicInfoInputProps {
-   name: string; //useActionsState로 매핑하기 위한
-   text: string;
-   placeholder: string;
-   defaultValue?: string;
-   onValidChange: (name: string, value: boolean) => void;
-   validator: (value: string) => ValidationResult; // 유효성 검사 함수
-   onChange?: (e: React.ChangeEvent<HTMLInputElement>) => void;
-}
